Create Tangent Circle Checker.c
Define a struct called a circle, whose members are centerx (decimal), center (decimal), and radius (decimal). Write a function called tegetMi, which checks whether these circles are tangent to each other according to the information of 2 circles sent to it, together with a main function where the circle information is entered from the keyboard and the result is printed on the screen (Global variable should not be used). If the function circle is not 1 to each other then 0 should be returned.

And the result must be printed "The circles are tangent" or "The circles are not tangent".